refactor: moved folder creation logic to app initliazer
