+++ backend/src/main/java/lk/ijse/webpos/backend/api/servlet/OrderServlet.java
@@ -2,8 +2,6 @@ package lk.ijse.webpos.backend.api.servlet;
 
 import jakarta.json.bind.Jsonb;
 import jakarta.json.bind.JsonbBuilder;
-import jakarta.servlet.ServletConfig;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -11,7 +9,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import lk.ijse.webpos.backend.bo.BOFactory;
 import lk.ijse.webpos.backend.bo.custom.OrderBO;
 import lk.ijse.webpos.backend.dto.OrderDTO;
-import lk.ijse.webpos.backend.util.DirectoryUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
